rename package sentry to sentryclient
