+++ sentryclient/project_keys.go
@@ -1,4 +1,4 @@
-package sentry
+package sentryclient
 
 import (
 	"net/http"
